+++ CHANGELOG.md
@@ -1266,3 +1266,4 @@
 - [2026-05-18 17:03] fix(storage): fix memory leak by cancelling storage subscription
 - [2026-05-18 17:37] feat(auth): implement shimmer loading skeleton in auth
 - [2026-05-18 17:57] refactor(i18n): optimize data flow between domain and data layers
+- [2026-05-18 18:01] fix(network): prevent double tap duplicate action in network